Port user login and Gafaelfawr mock
This ports the User/AuthenticatedUser classes from Mobu as well as the
mock_gafaelfawr for testing.

The key change is that we're using httpx throughout and respx as the
HTTP mocking library. We're also making User and AuthenticatedUser
straight-up dataclasses since we'll move data parsing and coercion into
the API layer rather than the domain/services.