Initial port of JupyterClient and mock
This is ported from mobu and modified to work with httpx, websockets,
and respx instead of aiohttp and aioresponses.

Both the JupyterClient and the JupyterLabSession is now a contexts so
that they automatically close their connections.

The JupyterLabSession is now where python execution happens, rather than
in the Jupyter Client, as done in mobu.