wip: Add testing code for Jupyter websockets
Mobu has code for mocking a websocket connection, thereby allowing tests
to exercise the code for running python cells in the JupyterLab kernel.
In the port from aiohttp to websockets, I've tried to also port that
mock, but haven't been able to get this to work yet. Given that we can't
exercise JupyterClient.open_websocket_session yet.

Instead of this mock I've been able to demonstrate that
JupyterLabSession works in real usage. It would be nice to get this to
work, but it may not provide much immediate value.