Implement an adapter for arq
This replaces the prototype arq_dependency so that instead of directly
supplying an ArqRedis instance, it instead provides a instance of a subclass
of ArqQueue. This adapter provides an interface to arq that is
implemented by either a redis-backed queue, or an in-memory mock of the
queue that can be useful for testing.