Enable JupyterClient to work outside a context
An async context manager doesn't work well for the arq worker
application, where there are separate start up and shutdown methods.
This commit makes it possible to use JupyterClient outside an async
context manager; the client is automatically opened by accessing either
the http_client or cachemachine properties. The user needs to remember
to call close in this case.