Provide way to try the next JupyterLab identity
If a worker exits and the JupyterLab pod does not successfully close, it
becomes orphaned. If a new worker picks up the identity of the orphaned
JupyterLab pod, its start-up sequence will fail. The
IdentityManager.get_next_identity() method provides a way for the worker
to try the next available identity in that circumstance.