Add JupyterLab keep alive cron job
The work on the JupyterLab culler means that noteburst's JL pods are
also being culled. At the moment, the JL culler is running at a very
high frequency (1 minute of no network activity and 5 minutes of no
network activity). This arq cron function attempts to exercise the JL
kernel every 30 seconds. This can be backed off when the JL culler is
relaxed to more reasonable frequencies.