Treat work identity uid as optional
Since Gafaelfawr can acquire a UID for us in some environments, we don't
want to pre-configure a worker identity uid in those cases. Thus the
identity YAML model can treat a missing uid field as `None`. When
getting a token, the uid is left out of the request.

With this change, the worker's username is now used in the identity
lock, rather than the uid.