Add configuration for worker keep-alive
Previously the worker keep_alive function, which periodically exercises
a worker's connection to the JupyterLab pod's kernel to prevent culling,
ran every 30 seconds to default an aggressive culler setting on
data-dev. Normally we don't want to run this fast, and in some cases we
don't want to run keep_alive at all for debugging.

This worker_keepalive setting can be "disabled", "fast", or "normal" to
disable or tune the frequency of the keep alive function.