Add a timeout field to POST notebook
This sets a timeout on the nbexec call to the JupyterLab server
extension for notebook execution. It raises a NbexecTaskTimeout
exception in the arq worker, which should result in an errored state for
the notebook execution job. We might need to fine tune the exception
handling so that the result output in the API is useful.