Improve nublado spawning flow in worker start-up
Overall clean up the start up code for spawning a JupyterLab session.
The spawn attempts now happen in a NubladoPod class so that we can
refactor a lot of the low-level code out of the statup function. Being a
dataclass, we can still retrieve the nublado client and logger with
bound context data.

Start sending spawn errors to Sentry with enriched data. We're currently
sending a mixture of Sentry and Slack-native exceptions, so this is
something we'll resolve in the future.

In the spawning loop, we now specifically catch the JupyterWebError,
which we weren't before, but comes up if we attempt to spawn a pod that
already exists.