Restart the worker if any issue from the lab
Before we were only catching specific errors to cause the worker to
restart and I think that was causing us to miss instances where the
JupyterLab pod was no longer working, but keep alive continued to pass.
Now we catch log, and ultimately call sys.exit on all types of
exceptions.