Avoid creating a lab session
This adds a execute_notebook method to the NubladoPod class that
replicates the functionality of
JupyterLabSession.run_notebook_via_rsp_extension, but without having to
set up the JupyterLabSession.

I found that creating the JupyterLabSession itself was a source of
unreliability (rubin.nublado.client.exceptions.JupyterWebError:
ReadTimeout when calling user/{username}/api/sessions). Using the HTTP
extensions doesn't require a websocket session, so this is an attempt to
see if we can improve reliability by skipping that setup.