Add basic test for keep_alive worker function
Use the Jupyter and Repertoire mocks to add a basic test of the
`keep_alive` worker function.

This is not yet a proper test because it doesn't run the worker
context setup and bypasses some of the Nublado code. A proper test
will come after there is a standard Gafaelfawr client and mock
available.